move model and local interfaces
