+++ src/main/java/org/mae/twg/backend/models/Local.java
@@ -1,6 +1,5 @@
-package org.mae.twg.backend.models.travel.localization;
+package org.mae.twg.backend.models;
 
-import org.mae.twg.backend.models.travel.Model;
 import org.mae.twg.backend.models.travel.enums.Localization;
 
 public interface Local {
