+++ src/main/java/org/mae/twg/backend/models/Model.java
@@ -1,6 +1,4 @@
-package org.mae.twg.backend.models.travel;
-
-import org.mae.twg.backend.models.travel.localization.Local;
+package org.mae.twg.backend.models;
 
 import java.util.List;
