+++ src/main/java/org/mae/twg/backend/models/travel/Tour.java
@@ -2,6 +2,7 @@ package org.mae.twg.backend.models.travel;
 
 import jakarta.persistence.*;
 import lombok.*;
+import org.mae.twg.backend.models.Model;
 import org.mae.twg.backend.models.business.Agency;
 import org.mae.twg.backend.models.travel.enums.TourType;
 import org.mae.twg.backend.models.travel.localization.TourLocal;
