+++ src/main/java/org/mae/twg/backend/models/travel/localization/HotelLocal.java
@@ -2,8 +2,9 @@ package org.mae.twg.backend.models.travel.localization;
 
 import jakarta.persistence.*;
 import lombok.*;
+import org.mae.twg.backend.models.Local;
 import org.mae.twg.backend.models.travel.Hotel;
-import org.mae.twg.backend.models.travel.Model;
+import org.mae.twg.backend.models.Model;
 import org.mae.twg.backend.models.travel.enums.Localization;
 
 @Entity
@@ -13,7 +14,7 @@ import org.mae.twg.backend.models.travel.enums.Localization;
 @AllArgsConstructor
 @NoArgsConstructor
 @Table(name = "hotel_local")
-public class HotelLocal implements Local{
+public class HotelLocal implements Local {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
