+++ src/main/java/org/mae/twg/backend/models/travel/localization/PropertyLocal.java
@@ -2,8 +2,9 @@ package org.mae.twg.backend.models.travel.localization;
 
 import jakarta.persistence.*;
 import lombok.*;
+import org.mae.twg.backend.models.Local;
 import org.mae.twg.backend.models.travel.Property;
-import org.mae.twg.backend.models.travel.Model;
+import org.mae.twg.backend.models.Model;
 import org.mae.twg.backend.models.travel.enums.Localization;
 
 @Entity
@@ -13,7 +14,7 @@ import org.mae.twg.backend.models.travel.enums.Localization;
 @AllArgsConstructor
 @NoArgsConstructor
 @Table(name = "property_local")
-public class PropertyLocal implements Local{
+public class PropertyLocal implements Local {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
