+++ src/main/java/org/mae/twg/backend/models/travel/localization/TourLocal.java
@@ -2,8 +2,9 @@ package org.mae.twg.backend.models.travel.localization;
 
 import jakarta.persistence.*;
 import lombok.*;
+import org.mae.twg.backend.models.Local;
 import org.mae.twg.backend.models.travel.Tour;
-import org.mae.twg.backend.models.travel.Model;
+import org.mae.twg.backend.models.Model;
 import org.mae.twg.backend.models.travel.enums.Localization;
 
 @Entity
@@ -13,7 +14,7 @@ import org.mae.twg.backend.models.travel.enums.Localization;
 @AllArgsConstructor
 @NoArgsConstructor
 @Table(name = "tour_local")
-public class TourLocal implements Local{
+public class TourLocal implements Local {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
